Remove redundant isTruthy and extract merge method into extension
