+++ Sources/xcstrings-tool/Generate.swift
@@ -60,7 +60,7 @@ struct Generate: ParsableCommand {
         }
 
         // Merge the resources together, ensure that they are uniquely keyed and sorted
-        let resources = try mergeAndEnsureUnique(results)
+        let resources = try StringExtractor.mergeAndEnsureUnique(results)
 
         // Generate the associated Swift source
         let source = StringGenerator.generateSource(
@@ -103,27 +103,5 @@ struct Generate: ParsableCommand {
 
     var resolvedAccessLevel: AccessLevel {
         .resolveFromEnvironment(or: accessLevel) ?? .internal
-    }
-
-    func mergeAndEnsureUnique(_ results: [StringExtractor.Result]) throws -> [Resource] {
-        if results.isEmpty { return [] }
-        if results.count == 1 { return results[0].resources }
-
-        let resources = results.flatMap { $0.resources }
-
-        let keyed = Dictionary(grouping: resources, by: \.key)
-        let conflicts = keyed.filter { $0.value.count > 1 }
-
-        guard conflicts.isEmpty else {
-            let conflicts = Array(conflicts.map(\.key)).formatted()
-            throw Diagnostic(
-                severity: .error,
-                message: """
-                Conflicting keys were found within the inputs: \(conflicts)
-                """
-            )
-        }
-
-        return resources.sorted()
     }
 }
