--- Sources/xcstrings-tool/Utilities/String+Truthy.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-extension String {
-    var isTruthy: Bool {
-        ["YES", "TRUE", "1"].contains(self.uppercased())
-    }
-}
